+++ packages/js-auth/src/index.ts
@@ -1,11 +1,7 @@
 export { core } from './core.js'
 export { provisioning } from './provisioning.js'
 
-export type * from './types/index.js'
-export type * from './types/authorizationCode.js'
-export type * from './types/clientCredentials.js'
-export type * from './types/password.js'
-export type * from './types/refreshToken.js'
+export type { GrantType, TOptions, TReturn } from './types/index.js'
 
 export type {
   TProvisioningOptions,
