+++ packages/js-auth/src/types/authorizationCode.ts
@@ -1,11 +1,11 @@
-import type { TBaseOptions } from '#types/index.js'
+import type { TBaseOptions } from '#types/base.js'
 import type { TPasswordReturn } from './password.js'
 
 /**
  * The authorization code grant type is used by clients to exchange an authorization code for an access token.
  * @see https://docs.commercelayer.io/core/authentication/authorization-code#getting-an-access-token
  */
-export interface TAuthorizationCode extends TBaseOptions {
+export interface TAuthorizationCodeOptions extends TBaseOptions {
   code: string
   redirectUri: string
   clientSecret: string
