+++ packages/js-auth/src/types/clientCredentials.ts
@@ -1,9 +1,9 @@
-import type { TBaseOptions } from '#types/index.js'
+import type { TBaseOptions } from '#types/base.js'
 
 /**
  * The client credentials grant type is used by clients to obtain an access token outside of the context of a user.
  * @see https://docs.commercelayer.io/core/authentication/client-credentials
  */
-export interface TClientCredentials extends TBaseOptions {
+export interface TClientCredentialsOptions extends TBaseOptions {
   clientSecret?: string
 }
