+++ packages/js-auth/src/types/password.ts
@@ -1,10 +1,10 @@
-import type { TBaseOptions, TBaseReturn } from '#types/index.js'
+import type { TBaseOptions, TBaseReturn } from '#types/base.js'
 
 /**
  * The password grant type is used by first-party clients to exchange a user's credentials for an access token.
  * @see https://docs.commercelayer.io/core/authentication/password
  */
-export interface TPassword extends TBaseOptions {
+export interface TPasswordOptions extends TBaseOptions {
   username: string
   password: string
 }
